New scene +Float textures +highp on shaders

+++ src/shaders/postprocessing.frag
@@ -1,4 +1,4 @@
-precision mediump float;
+precision highp float;
 
 varying vec2 fragUV;
 uniform sampler2D texture;
@@ -23,7 +23,7 @@ void make_kernel(inout vec4 n[9], sampler2D tex, vec2 coord) {
 
 void main(void) {
   vec4 n[9];
-	make_kernel( n, texture, fragUV.st );
+	make_kernel(n, texture, fragUV.st);
   vec4 sobel_edge_h = n[2] + (2.0*n[5]) + n[8] - (n[0] + (2.0*n[3]) + n[6]);
   vec4 sobel_edge_v = n[0] + (2.0*n[1]) + n[2] - (n[6] + (2.0*n[7]) + n[8]);
   vec4 sobel = sqrt((sobel_edge_h * sobel_edge_h) + (sobel_edge_v * sobel_edge_v));
